+++ src/gaet/core.py
@@ -66,7 +66,7 @@ from pathlib import Path
 
 from typing import Any, Dict, List, Optional, Tuple
 
-VERSION = "3.0.0"
+VERSION = "3.0.1"
 
 NAME = "gaet"